+++ lib/src/html_renderer.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:collection';
+import 'dart:convert';
 
 import 'ast.dart';
 import 'block_parser.dart';
@@ -17,8 +18,8 @@ String markdownToHtml(String markdown,
     ExtensionSet extensionSet,
     Resolver linkResolver,
     Resolver imageLinkResolver,
-    bool inlineOnly: false}) {
-  var document = new Document(
+    bool inlineOnly = false}) {
+  var document = Document(
       blockSyntaxes: blockSyntaxes,
       inlineSyntaxes: inlineSyntaxes,
       extensionSet: extensionSet,
@@ -34,50 +35,79 @@ String markdownToHtml(String markdown,
 }
 
 /// Renders [nodes] to HTML.
-String renderToHtml(List<Node> nodes) => new HtmlRenderer().render(nodes);
+String renderToHtml(List<Node> nodes) => HtmlRenderer().render(nodes);
+
+const _blockTags = [
+  'blockquote',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'hr',
+  'li',
+  'ol',
+  'p',
+  'pre',
+  'ul',
+];
 
 /// Translates a parsed AST to HTML.
 class HtmlRenderer implements NodeVisitor {
-  static final _blockTags = new RegExp('blockquote|h1|h2|h3|h4|h5|h6|hr|p|pre');
-
   StringBuffer buffer;
   Set<String> uniqueIds;
+
+  final _elementStack = <Element>[];
+  String _lastVisitedTag;
 
   HtmlRenderer();
 
   String render(List<Node> nodes) {
-    buffer = new StringBuffer();
-    uniqueIds = new LinkedHashSet<String>();
+    buffer = StringBuffer();
+    uniqueIds = LinkedHashSet<String>();
 
-    for (final node in nodes) node.accept(this);
+    for (final node in nodes) {
+      node.accept(this);
+    }
 
     return buffer.toString();
   }
 
   void visitText(Text text) {
-    buffer.write(text.text);
+    var content = text.text;
+    if (const ['p', 'li'].contains(_lastVisitedTag)) {
+      var lines = LineSplitter.split(content);
+      content = (content.contains('<pre>'))
+          ? lines.join('\n')
+          : lines.map((line) => line.trimLeft()).join('\n');
+      if (text.text.endsWith('\n')) {
+        content = '$content\n';
+      }
+    }
+    buffer.write(content);
+
+    _lastVisitedTag = null;
   }
 
   bool visitElementBefore(Element element) {
     // Hackish. Separate block-level elements with newlines.
-    if (buffer.isNotEmpty && _blockTags.firstMatch(element.tag) != null) {
-      buffer.write('\n');
+    if (buffer.isNotEmpty && _blockTags.contains(element.tag)) {
+      buffer.writeln();
     }
 
     buffer.write('<${element.tag}');
 
-    // Sort the keys so that we generate stable output.
-    var attributeNames = element.attributes.keys.toList();
-    attributeNames.sort((a, b) => a.compareTo(b));
-
-    for (var name in attributeNames) {
-      buffer.write(' $name="${element.attributes[name]}"');
+    for (var entry in element.attributes.entries) {
+      buffer.write(' ${entry.key}="${entry.value}"');
     }
 
     // attach header anchor ids generated from text
     if (element.generatedId != null) {
       buffer.write(' id="${uniquifyId(element.generatedId)}"');
     }
+
+    _lastVisitedTag = element.tag;
 
     if (element.isEmpty) {
       // Empty element like <hr/>.
@@ -89,13 +119,26 @@ class HtmlRenderer implements NodeVisitor {
 
       return false;
     } else {
+      _elementStack.add(element);
       buffer.write('>');
       return true;
     }
   }
 
   void visitElementAfter(Element element) {
+    assert(identical(_elementStack.last, element));
+
+    if (element.children != null &&
+        element.children.isNotEmpty &&
+        _blockTags.contains(_lastVisitedTag) &&
+        _blockTags.contains(element.tag)) {
+      buffer.writeln();
+    } else if (element.tag == 'blockquote') {
+      buffer.writeln();
+    }
     buffer.write('</${element.tag}>');
+
+    _lastVisitedTag = _elementStack.removeLast().tag;
   }
 
   /// Uniquifies an id generated from text.