+++ lib/src/util.dart
@@ -5,11 +5,14 @@ import 'package:charcode/charcode.dart';
 String escapeHtml(String html) =>
     const HtmlEscape(HtmlEscapeMode.element).convert(html);
 
+String escapeHtmlAttribute(String text) =>
+    const HtmlEscape(HtmlEscapeMode.attribute).convert(text);
+
 // Escape the contents of [value], so that it may be used as an HTML attribute.
 
 // Based on http://spec.commonmark.org/0.28/#backslash-escapes.
 String escapeAttribute(String value) {
-  var result = new StringBuffer();
+  var result = StringBuffer();
   int ch;
   for (var i = 0; i < value.codeUnits.length; i++) {
     ch = value.codeUnitAt(i);