+++ test/markdown_test.dart
@@ -20,9 +20,26 @@ void main() {
   testFile('extensions/tables.unit', blockSyntaxes: [const TableSyntax()]);
 
   // Inline syntax extensions
-  testFile('extensions/emojis.unit', inlineSyntaxes: [new EmojiSyntax()]);
-  testFile('extensions/inline_html.unit',
-      inlineSyntaxes: [new InlineHtmlSyntax()]);
+  testFile('extensions/emojis.unit', inlineSyntaxes: [EmojiSyntax()]);
+  testFile('extensions/inline_html.unit', inlineSyntaxes: [InlineHtmlSyntax()]);
+
+  testDirectory('common_mark');
+  testDirectory('gfm', extensionSet: ExtensionSet.gitHubFlavored);
+
+  group('Corner cases', () {
+    validateCore('Incorrect Links', '''
+5 Ethernet ([Music](
+''', '''
+<p>5 Ethernet ([Music](</p>
+''');
+
+    validateCore('Escaping code block language', '''
+```"/><a/href="url">arbitrary_html</a>
+```
+''', '''
+<pre><code class="language-&quot;/&gt;&lt;a/href=&quot;url&quot;&gt;arbitrary_html&lt;/a&gt;"></code></pre>
+''');
+  });
 
   group('Corner cases', () {
     validateCore(
@@ -37,7 +54,7 @@ void main() {
 
   group('Resolver', () {
     Node nyanResolver(String text, [_]) =>
-        text.isEmpty ? null : new Text('~=[,,_${text}_,,]:3');
+        text.isEmpty ? null : Text('~=[,,_${text}_,,]:3');
     validateCore(
         'simple link resolver',
         '''
@@ -100,7 +117,7 @@ resolve [[]] thing
   });
 
   group('Custom inline syntax', () {
-    var nyanSyntax = <InlineSyntax>[new TextSyntax('nyan', sub: '~=[,,_,,]:3')];
+    var nyanSyntax = <InlineSyntax>[TextSyntax('nyan', sub: '~=[,,_,,]:3')];
     validateCore(
         'simple inline syntax',
         '''
@@ -112,7 +129,7 @@ nyan''',
     validateCore('dart custom links', 'links [are<foo>] awesome',
         '<p>links <a>are&lt;foo></a> awesome</p>\n',
         linkResolver: (String text, [_]) =>
-            new Element.text('a', text.replaceAll('<', '&lt;')));
+            Element.text('a', text.replaceAll('<', '&lt;')));
 
     // TODO(amouravski): need more tests here for custom syntaxes, as some
     // things are not quite working properly. The regexps are sometime a little