+++ tool/stats.dart
@@ -5,12 +5,16 @@ import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:collection/collection.dart';
+import 'package:expected_output/expected_output.dart';
 import 'package:path/path.dart' as p;
 
 import 'stats_lib.dart';
 
-Future main(List<String> args) async {
-  final parser = new ArgParser()
+final _configs =
+    List<Config>.unmodifiable([Config.commonMarkConfig, Config.gfmConfig]);
+
+Future<void> main(List<String> args) async {
+  final parser = ArgParser()
     ..addOption('section',
         help: 'Restrict tests to one section, provided after the option.')
     ..addFlag('raw',
@@ -27,9 +31,7 @@ Future main(List<String> args) async {
         defaultsTo: false,
         help: 'Print details for "loose" matches.',
         negatable: false)
-    ..addOption('flavor',
-        allowed: [Config.commonMarkConfig.prefix, Config.gfmConfig.prefix],
-        defaultsTo: Config.commonMarkConfig.prefix)
+    ..addOption('flavor', allowed: _configs.map((c) => c.prefix))
     ..addFlag('help', defaultsTo: false, negatable: false);
 
   ArgResults options;
@@ -61,37 +63,83 @@ Future main(List<String> args) async {
     return;
   }
 
-  final testPrefix = options['flavor'] as String;
-
-  Config config;
-  switch (testPrefix) {
-    case 'gfm':
-      config = Config.gfmConfig;
-      break;
-    case 'common_mark':
-      config = Config.commonMarkConfig;
-      break;
-    default:
-      throw new ArgumentError('Does not support `$testPrefix`.');
+  var testPrefix = options['flavor'] as String;
+  if (!updateFiles) {
+    testPrefix = _configs.first.prefix;
   }
+
+  final testPrefixes =
+      testPrefix == null ? _configs.map((c) => c.prefix) : <String>[testPrefix];
+
+  for (var testPrefix in testPrefixes) {
+    await _processConfig(testPrefix, raw, updateFiles, verbose,
+        specifiedSection, verboseLooseMatch);
+  }
+}
+
+final _sectionNameReplace = RegExp('[ \\)\\(]+');
+
+String _unitOutput(Iterable<DataCase> cases) => cases.map((dataCase) => '''
+>>> ${dataCase.front_matter}
+${dataCase.input}<<<
+${dataCase.expectedOutput}''').join();
+
+/// Set this to `true` and rerun `--update-files` to ease finding easy strict
+/// fixes.
+const _improveStrict = false;
+
+Future<void> _processConfig(
+  String testPrefix,
+  bool raw,
+  bool updateFiles,
+  bool verbose,
+  String specifiedSection,
+  bool verboseLooseMatch,
+) async {
+  final config = _configs.singleWhere((c) => c.prefix == testPrefix);
 
   var sections = loadCommonMarkSections(testPrefix);
 
-  var scores = new SplayTreeMap<String, SplayTreeMap<int, CompareLevel>>(
+  var scores = SplayTreeMap<String, SplayTreeMap<int, CompareLevel>>(
       compareAsciiLowerCaseNatural);
 
-  sections.forEach((section, examples) {
-    if (specifiedSection != null && section != specifiedSection) {
-      return;
+  for (var entry in sections.entries) {
+    if (specifiedSection != null && entry.key != specifiedSection) {
+      continue;
     }
-    for (var e in examples) {
-      var nestedMap = scores.putIfAbsent(
-          section, () => new SplayTreeMap<int, CompareLevel>());
 
-      nestedMap[e.example] = compareResult(config, e,
+    final units = <DataCase>[];
+
+    for (var e in entry.value) {
+      final result = compareResult(config, e,
           verboseFail: verbose, verboseLooseMatch: verboseLooseMatch);
+
+      units.add(DataCase(
+        front_matter: result.testCase.toString(),
+        input: result.testCase.markdown,
+        expectedOutput:
+            (_improveStrict && result.compareLevel == CompareLevel.loose)
+                ? result.testCase.html
+                : result.result,
+      ));
+
+      var nestedMap = scores.putIfAbsent(
+          entry.key, () => SplayTreeMap<int, CompareLevel>());
+      nestedMap[e.example] = result.compareLevel;
     }
-  });
+
+    if (updateFiles && units.isNotEmpty) {
+      var fileName =
+          entry.key.toLowerCase().replaceAll(_sectionNameReplace, '_');
+      while (fileName.endsWith('_')) {
+        fileName = fileName.substring(0, fileName.length - 1);
+      }
+      fileName = '$fileName.unit';
+      File(p.join('test', testPrefix, fileName))
+        ..createSync(recursive: true)
+        ..writeAsStringSync(_unitOutput(units));
+    }
+  }
 
   if (raw || updateFiles) {
     await _printRaw(testPrefix, scores, updateFiles);
@@ -102,7 +150,7 @@ Future main(List<String> args) async {
   }
 }
 
-Object _convert(obj) {
+Object _convert(Object obj) {
   if (obj is CompareLevel) {
     switch (obj) {
       case CompareLevel.strict:
@@ -114,11 +162,11 @@ Object _convert(obj) {
       case CompareLevel.loose:
         return 'loose';
       default:
-        throw new ArgumentError("`$obj` is unknown.");
+        throw ArgumentError("`$obj` is unknown.");
     }
   }
   if (obj is Map) {
-    var map = {};
+    var map = <String, Object>{};
     obj.forEach((k, v) {
       var newKey = k.toString();
       map[newKey] = v;
@@ -128,7 +176,8 @@ Object _convert(obj) {
   return obj;
 }
 
-Future _printRaw(String testPrefix, Map scores, bool updateFiles) async {
+Future<void> _printRaw(String testPrefix,
+    Map<String, Map<int, CompareLevel>> scores, bool updateFiles) async {
   IOSink sink;
   if (updateFiles) {
     var file = getStatsFile(testPrefix);
@@ -151,20 +200,24 @@ Future _printRaw(String testPrefix, Map scores, bool updateFiles) async {
   await sink.close();
 }
 
-Future _printFriendly(
+String _pct(int value, int total, String section) =>
+    '${value.toString().padLeft(4)} '
+    'of ${total.toString().padLeft(4)} '
+    '– ${(100 * value / total).toStringAsFixed(1).padLeft(5)}%  $section';
+
+Future<void> _printFriendly(
     String testPrefix,
     SplayTreeMap<String, SplayTreeMap<int, CompareLevel>> scores,
     bool updateFiles) async {
-  const countWidth = 4;
-
   var totalValid = 0;
+  var totalStrict = 0;
   var totalExamples = 0;
 
   IOSink sink;
   if (updateFiles) {
     var path = p.join(toolDir, '${testPrefix}_stats.txt');
     print('Updating $path');
-    var file = new File(path);
+    var file = File(path);
     sink = file.openWrite();
   } else {
     sink = stdout;
@@ -182,20 +235,14 @@ Future _printFriendly(
 
     var sectionValidCount = sectionStrictCount + sectionLooseCount;
 
+    totalStrict += sectionStrictCount;
     totalValid += sectionValidCount;
 
-    var pct = (100 * sectionValidCount / total).toStringAsFixed(1).padLeft(5);
-
-    sink.writeln('${sectionValidCount.toString().padLeft(countWidth)} '
-        'of ${total.toString().padLeft(countWidth)} '
-        '– $pct%  $section');
+    sink.writeln(_pct(sectionValidCount, total, section));
   });
 
-  var pct = (100 * totalValid / totalExamples).toStringAsFixed(1).padLeft(5);
-
-  sink.writeln('${totalValid.toString().padLeft(countWidth)} '
-      'of ${totalExamples.toString().padLeft(countWidth)} '
-      '– $pct%  TOTAL');
+  sink.writeln(_pct(totalValid, totalExamples, 'TOTAL'));
+  sink.writeln(_pct(totalStrict, totalValid, 'TOTAL Strict'));
 
   await sink.flush();
   await sink.close();