+++ tool/update_blns.dart
@@ -7,25 +7,28 @@ final _blnsJsonRawUrl =
 final _blnsFilePath = 'test/blns.dart';
 
 Future<Null> main() async {
-  var client = new HttpClient();
+  var client = HttpClient();
   List<String> json;
   try {
     var request = await client.getUrl(Uri.parse(_blnsJsonRawUrl));
     var response = await request.close();
-    json = (jsonDecode(await response.transform(utf8.decoder).join('')) as List)
+    json = (jsonDecode(await response
+            .cast<List<int>>()
+            .transform(utf8.decoder)
+            .join('')) as List)
         .cast<String>();
   } finally {
     client.close();
   }
-  var blnsContent = new StringBuffer('''
+  var blnsContent = StringBuffer('''
 // GENERATED FILE. DO NOT EDIT.
 //
 // This file was generated from big-list-of-naughty-strings's JSON file:
 // $_blnsJsonRawUrl
-// at ${new DateTime.now()} by the script, tool/update_blns.dart.
+// at ${DateTime.now()} by the script, tool/update_blns.dart.
 
 ''');
-  blnsContent.writeln('const blns = const <String>[');
+  blnsContent.writeln('const blns = <String>[');
   for (var str in json) {
     var escaped = str
         .replaceAll(r'\', r'\\')
@@ -34,5 +37,5 @@ Future<Null> main() async {
     blnsContent.writeln("  '$escaped',");
   }
   blnsContent.writeln('];');
-  new File(_blnsFilePath)..writeAsStringSync(blnsContent.toString());
+  File(_blnsFilePath)..writeAsStringSync(blnsContent.toString());
 }