+++ tool/update_emojis.dart
@@ -7,21 +7,22 @@ final _emojisJsonRawUrl =
 final _emojisFilePath = 'lib/src/emojis.dart';
 
 Future<Null> main() async {
-  var client = new HttpClient();
+  var client = HttpClient();
   var request = await client.getUrl(Uri.parse(_emojisJsonRawUrl));
   var response = await request.close();
-  var json = jsonDecode(await response.transform(utf8.decoder).join(''))
-      .map((alias, info) => new MapEntry(alias, info.cast<String, dynamic>()))
+  var json = jsonDecode(
+          await response.cast<List<int>>().transform(utf8.decoder).join(''))
+      .map((alias, info) => MapEntry(alias, info.cast<String, dynamic>()))
       .cast<String, Map<String, dynamic>>();
-  var emojisContent = new StringBuffer('''
+  var emojisContent = StringBuffer('''
 // GENERATED FILE. DO NOT EDIT.
 //
 // This file was generated from emojilib's emoji data file:
 // $_emojisJsonRawUrl
-// at ${new DateTime.now()} by the script, tool/update_emojis.dart.
+// at ${DateTime.now()} by the script, tool/update_emojis.dart.
 
 ''');
-  emojisContent.writeln('const emojis = const <String, String>{');
+  emojisContent.writeln('const emojis = <String, String>{');
   var emojiCount = 0;
   var ignored = <String>[];
   json.forEach((String alias, Map<String, dynamic> info) {
@@ -33,7 +34,7 @@ Future<Null> main() async {
     }
   });
   emojisContent.writeln('};');
-  new File(_emojisFilePath)..writeAsStringSync(emojisContent.toString());
+  File(_emojisFilePath)..writeAsStringSync(emojisContent.toString());
   print('Wrote data to $_emojisFilePath for $emojiCount emojis, '
       'ignoring ${ignored.length}: ${ignored.join(', ')}.');
   exit(0);